07.Python_Web/07.01.Django_Basics/07.01.01.Labs/forumApp/requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.1
django-crispy-forms==2.3
Markdown==3.7
psycopg2-binary==2.9.9
sqlparse==0.5.1
